Cosmetics

+++ radio/src/gui/480x272/model_logical_switches.cpp
@@ -102,7 +102,7 @@ bool menuModelLogicalSwitches(event_t event)
     POPUP_MENU_START(onLogicalSwitchesMenu);
   }
 
-  for (int i=0; i<NUM_BODY_LINES; ++i) {
+  for (uint8_t i=0; i<NUM_BODY_LINES; ++i) {
     coord_t y = MENU_CONTENT_TOP + i*FH;
     int k = i+menuVerticalOffset;
     LcdFlags attr = (menuVerticalPosition==k ? ((s_editMode>0) ? BLINK|INVERS : INVERS)  : 0);
